+++ src/formatFileUrl.js
@@ -1,0 +1,6 @@
+function formatFileUrl(pathToFile, fileName) {
+  const name = fileName.replace(/.md$/, '');
+  return (pathToFile) ? `${pathToFile}/${name}` : name;
+}
+
+module.exports = formatFileUrl