+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 - [x] Create resources directory tree and add blank README.md files in each
 - [ ] Fill in each README.md file under the resources tree with description the subject of the directory
 - [ ] Test travis configuration with a test index.md and/or index.html in the public directory
-- [ ] Add a test.md in public directory
-- [ ] Automatically generate an index with a link to test.md
+- [x] Add a test.md in public directory
+- [x] Automatically generate an index with a link to test.md